debug

+++ llava/train/train.py
@@ -104,7 +104,7 @@ class TrainingArguments(transformers.TrainingArguments):
     )
     lora_enable: bool = False
     lora_r: int = 64
-    lora_alpha: int = 16
+    lora_alpha: float = 16
     lora_dropout: float = 0.05
     lora_weight_path: str = ""
     lora_bias: str = "none"
